365. Water and Jug Problem
Commonly, there are two algorithm for this Problem. One is GCD, the
other is simulation. But for simulation, memory limit exceeded is always
thrown. The reason is, for judge false case, a queue or dictionary must
be maintained. For corner case, the queue or dictonary may be extended
more large, which will lead Memory Limit Extended error.